The first working AttendanceTimer program
Tired of checking the Windows EventViewer for you login time?
This tool will probably ease your pain. It checks for your first
login of the day and computes your logout time based on a 9.6hrs
a day working schedule. It will then display a countdown timer of
the working hours left so that you can monitor the time you left
for your work. Once it reaches 9.6 hrs, the timer will stop at
00:00:00 and prompts a message that your duty hours is done. Also,
the date, login and logout time will be displayed for additional info.